完整测试用例

+++ game-utils/src/main/java/com/wjybxx/fastjgame/util/ThreadUtils.java
@@ -80,7 +80,7 @@ public class ThreadUtils {
      *
      * @param sleepMillis 要睡眠的时间(毫秒)
      */
-    public static void sleepQuietly(int sleepMillis) {
+    public static void sleepQuietly(long sleepMillis) {
         try {
             Thread.sleep(sleepMillis);
         } catch (InterruptedException ignore) {
